+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## [1.0.0] - 20/July/2020
+* Initial release Function available: doesOverlap() and isPointInPolygon().
 ## [1.1.0] - 20/July/2020
-* Initial release Function available: doesOverlap() and isPointInPolygon()
+* Minor Tweaks.
+## [1.2.0] - 20/July/2020
+* Added isConvexPolygon() Function and the ability to feed concave polygons too.
+